# app/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/suggestions/merge": {
      "post": {
        "operationId": "SuggestionsController_merge",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Результат слияния"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Слияние предложений (sourceId → targetId)",
        "tags": [
          "Suggestions"
        ]
      }
    },
...
    "/api/v1/suggestions/bulk/delete": {
      "post": {
        "operationId": "BulkController_bulkDelete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно удалено"
          },
          "207": {
            "description": "Частичный успех (часть не удалена)"
          }
        },
        "summary": "Массовое удаление",
        "tags": [
          "Bulk"
        ]
      }
    },
    "/api/v1/suggestions/bulk/status": {
      "post": {
        "operationId": "BulkController_bulkUpdateStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно обновлено"
          },
          "207": {
            "description": "Частичный успех"
          }
        },
        "summary": "Массовое изменение статуса",
        "tags": [
          "Bulk"
        ]
      }
    },
    "/api/v1/suggestions/bulk/merge": {
      "post": {
        "operationId": "BulkController_bulkMerge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkMergeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно слито"
          },
          "207": {
            "description": "Частичный успех"
          }
        },
        "summary": "Массовое слияние",
        "tags": [
          "Bulk"
        ]
      }
    },
...
  },
  "info": {
    "title": "Suggestions API",
    "description": "API для управления предложениями (suggestions)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BulkDeleteDto": {
        "type": "object",
        "properties": {}
      },
      "BulkUpdateStatusDto": {
        "type": "object",
        "properties": {}
      },
      "BulkMergeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}