# app/src/modules/suggestions/dto/bulk-merge.dto.ts
import { ArrayNotEmpty, IsArray, IsString } from 'class-validator';

export class BulkMergeDto {
  @IsArray()
  @ArrayNotEmpty()
  @IsString({ each: true })
  sourceIds!: string[];

  @IsString()
  targetId!: string;
}